Fix unit tests and load deadlock
The load unit-test hanged indefinitely. Presumably, this is because of a
deadlock caused by starting a write transaction inside a read
transaction. From the boltdb docs:

> Read-only transactions and read-write transactions should not depend on
> one another and generally shouldn't be opened simultaneously in the
> same goroutine. This can cause a deadlock as the read-write transaction
> needs to periodically re-map the data file but it cannot do so while a
> read-only transaction is open.

In addition, the protobuf error message returned for a case of the load
unit test seems to have changed, so amended to the latest text.

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>